Make --dev null servers exit faster
The previous refactored version had the drawback that we'd wait for each server
to stop before moving on to the next one. This would have been very slow as
server count increased. Now poll for clients in all status files, so that we can
exit fast when no clients are connected to any server.

Signed-off-by: Samuli Seppänen <samuli.seppanen@gmail.com>