Latest --dev null tests
Use pid files to determine if clients are connected. This is way faster and more
reliable than relying on status files where UNDEF clients may linger on for a
long time after the clients have actually exited (despite
--explicit-exit-notify). With this commit reliability in stress tests is about
99% based on 300 iterations. Further reliability improvements require the
clients to be more accurate in determining when the server is actually up and
able to process client connections.

Signed-off-by: Samuli Seppänen <samuli.seppanen@gmail.com>